Issue #45 Replay resume shall not cleanup state storage as regular replay

+++ libraries/plugins/account_history_rocksdb/account_history_rocksdb_plugin.cpp
@@ -1301,13 +1301,15 @@ void account_history_rocksdb_plugin::impl::prunePotentiallyTooOldItems(account_h
 
 void account_history_rocksdb_plugin::impl::on_pre_reindex(const hive::chain::reindex_notification& note)
 {
-   ilog("Received onReindexStart request, attempting to clean database storage.");
-
    shutdownDb();
    std::string strPath = _storagePath.string();
 
-   auto s = ::rocksdb::DestroyDB(strPath, ::rocksdb::Options());
-   checkStatus(s);
+   if( note.replay_clean )
+   {
+      ilog("Received onReindexStart request, attempting to clean database storage.");
+      auto s = ::rocksdb::DestroyDB(strPath, ::rocksdb::Options());
+      checkStatus(s);
+   }
 
    openDb();
